Move project status management to a single module
