+++ src/sync/index.ts
@@ -29,17 +29,12 @@ import {DEBUG_APP} from '../buildconfig';
 import {ProjectDataObject, ProjectMetaObject} from 'faims3-datamodel';
 import {
   data_dbs,
-  ExistingActiveDoc,
   ListingsObject,
   metadata_dbs,
   directory_db,
 } from './databases';
-import {
-  all_projects_updated,
-  createdProjects,
-  createdProjectsInterface,
-} from './state';
-import {logError} from '../logging';
+import {all_projects_updated} from './state';
+import {listenProject} from './projects';
 
 PouchDB.plugin(PouchDBFind);
 PouchDB.plugin(pouchdbDebug);
@@ -96,257 +91,6 @@ export async function waitForStateOnce(
   });
 }
 
-export async function getProject(
-  project_id: ProjectID
-): Promise<createdProjectsInterface> {
-  // Wait for all_projects_updated to possibly change before returning
-  // error/data DB if it's ready.
-  await waitForStateOnce(() => all_projects_updated);
-  if (project_id in data_dbs) {
-    return createdProjects[project_id];
-  } else {
-    throw `Project ${project_id} is not known`;
-  }
-}
-
-/**
- * Allows you to listen for changes from a Project's Data/Meta DBs or other
- * project info like if it's to be synced or not (from createdProjects)
- * This is a working alternative to getDataDB.changes
- * (as getDataDB.changes that may detach after updates to the owning listing
- * or the owning active DB, or if the sync is toggled on/off)
- *
- * @param project_id Full Project ID to listen on the DB for.
- * @param listener
- *     Called whenever the project you're listening on is available
- *     __Not necessarily has the data or metadata fully synced__
- *     But the data & metadata dbs will be in data_dbs, meta_dbs,
- *     and createdProjects.
- *     * meta_changed and data_changed events flow from
- *     the 'project_update' event in events.ts, and signal if the
- *     PouchDB databases have been recreated (and might need to
- *     be re-listened on)
- *     * error is available for the listener to call to asynchronously
- *     throw errors up to the error_listener. Use this instead of
- *     what you give into error_listener to ensure cleanup is done.
- *     * returns a destructor: This destructor is called when either
- *       * listenProject's destructor is called
- *       * Errors occur that mean we stop listening
- *       * The project info is *updated* (replaced will be true)
- *       * The project info is dropped (e.g. the user left)
- *     * Returning _'keep'_ changes behaviour: If this is a project info update,
- *       the destructor previously returned or kept from listener isn't run,
- *       and in fact, sticks around until next listener() (not returning keep)
- *       or other detach/error scenario.
- *     * Returning _'noop'_ returns a constructor doing nothing
- *       (This is not 'void' )
- * @param error_listener
- *     Called once at the first error condition.
- *     * All projects are synced, but project_id isn't a known project
- *     * errors in listener()
- *     * errors thrown asynchronously form listener
- *     * errors in the destructor from listener
- * @returns Detach function: call this to stop all changes
- */
-export function listenProject(
-  project_id: ProjectID,
-  // Listener, returning a destructor
-  // Listener receives an 'error' function to let it asynchronously throw errors.
-  // The destructor is called before a second listener is called
-  // but the destructor is optional
-  listener: (
-    value: createdProjectsInterface,
-    throw_error: (err: any) => void,
-    meta_changed: boolean,
-    data_changed: boolean
-  ) => 'keep' | 'noop' | ((replaced: boolean) => void),
-  error_listener: (value: unknown) => any
-): () => void {
-  if (DEBUG_APP) {
-    console.debug('listenProject starting');
-  }
-  // This is an array to allow it to be read/writeable from closures
-  const destructor: ['deleted' | 'initial' | ((replaced: boolean) => void)] = [
-    'initial',
-  ];
-
-  /* Set on a first error, to avoid multiple calls to error_listener */
-  const current_error: [null | {}] = [null];
-
-  /* Called when errors occur. Propagates to error_listener
-  but also runs cleanup */
-  const self_destruct = (err: unknown, detach = true) => {
-    if (DEBUG_APP) {
-      console.debug('listenProject running self_destruct');
-    }
-    // Only call error_listener once
-    if (current_error[0] === null) {
-      current_error[0] = (err as null | {}) ?? (Error('undefined error') as {});
-      try {
-        error_listener(err);
-      } catch (err: unknown) {
-        logError(err);
-        if (detach) {
-          detach_cb();
-        }
-        throw err; // Allow node to report as uncaught
-      }
-      if (detach) {
-        detach_cb();
-      }
-    }
-  };
-
-  const project_update_cb = (
-    type: ['update', createdProjectsInterface] | ['delete'] | ['create'],
-    meta_changed: boolean,
-    data_changed: boolean,
-    active: ExistingActiveDoc
-  ) => {
-    if (DEBUG_APP) {
-      console.debug('listenProject running project_update hook');
-    }
-    if (project_id === active._id) {
-      if (type[0] === 'delete') {
-        // Run destructor when the createdProjectsInterface object is deleted.
-        if (typeof destructor[0] !== 'function') {
-          logError(
-            'Non-fatal: listenProject destructor has gone ' +
-              "missing OR 'delete' event did not follow " +
-              "'update' or 'create' event"
-          );
-        } else {
-          destructor[0](false);
-        }
-        destructor[0] = 'deleted';
-      } else {
-        try {
-          const returned = listener(
-            createdProjects[active._id],
-            self_destruct,
-            meta_changed,
-            data_changed
-          );
-          if (returned !== 'keep') {
-            // If this is an update (destructor exists) then run destructor,
-            // and set the new destructor
-            if (typeof destructor[0] === 'function') {
-              if (type[0] !== 'update') {
-                console.warn(
-                  "Why is the destructor still around? either '" +
-                    `${type[0]} was triggered in the wrong place or some part` +
-                    " of this function didn't remove the destructor after use"
-                );
-              }
-              destructor[0](true);
-            }
-            if (returned === 'noop') {
-              // if the listener returned void
-              destructor[0] = () => {};
-            } else {
-              destructor[0] = returned;
-            }
-          }
-        } catch (err: unknown) {
-          self_destruct(err);
-        }
-      }
-    }
-  };
-
-  /*
-  All state is monitored because, just like getDataDB, when all projects are
-  known and the changes hasn't been set yet, the user has tried to listen on
-  a Data DB that doesn't exist.
-  */
-  const all_state_cb = () => {
-    if (DEBUG_APP) {
-      console.debug('listenProject running all_state hook');
-    }
-    if (all_projects_updated && destructor[0] === 'initial') {
-      self_destruct(Error(`Project ${project_id} is not known`));
-    } else if (all_projects_updated && destructor[0] === 'deleted') {
-      /*
-      In a flow that doesn't hit this warning:
-      1. The project is deleted, e.g. by the user leaving the project
-      2. project_update 'delete' event is emitted
-      3. __User of this function receives the delete event, and detaches
-           by calling the return of this function.__
-      3a. destructor is NOT CALLED with type: 'deleted'
-      4. Eventually (Or immediately after) all_state event is emitted with
-         all_projects_updated === true.
-      5. This function is NOT CALLED due to it being detached
-
-      As long as the user calls the detacher (Return of this function) between
-      a project_update 'delete' event and all_state is emitted, this warning is
-      not given.
-
-      Note: Event if 3a ('deleted') destructor is called before the user calls
-        the detacher, it still wouldn't error out because whilst the destructor
-        would run with 'deleted' and set to 'deleted', all_state would detach
-        by the user calling the detach function.
-      */
-      console.warn(
-        `Project ${project_id} did exist, was deleted, but a function` +
-          "listening to events on it's data DB didn't call the listener's " +
-          'detacher function at the right time (immediately after' +
-          'project_update event for the corresponding project id)'
-      );
-      // Allow the project to be undeleted & have listeners still work:
-      // So don't detach_cb here.
-    }
-  };
-
-  const detach_cb = () => {
-    if (DEBUG_APP) {
-      console.debug('listenProject running detach hook');
-    }
-    events.removeListener('project_update', project_update_cb);
-    events.removeListener('all_state', all_state_cb);
-    if (destructor[0] !== null && typeof destructor[0] === 'function') {
-      try {
-        destructor[0](false);
-      } catch (err: unknown) {
-        self_destruct(err, false);
-      }
-    }
-  };
-  if (DEBUG_APP) {
-    console.debug('listenProject created hooks');
-  }
-
-  // It's possible we'll never receive 'project_update' whilst listening (as it
-  // only gets called when the project information itself is changed, so invoke
-  // the callback if the project exists
-  const proj_info = createdProjects[project_id];
-  if (proj_info !== undefined) {
-    if (DEBUG_APP) {
-      console.debug('listenProject running initial callback');
-    }
-    try {
-      const returned = listener(proj_info, self_destruct, true, true);
-      if (returned !== 'keep') {
-        if (returned === 'noop') {
-          // if the listener returned void
-          destructor[0] = () => {};
-        } else {
-          destructor[0] = returned;
-        }
-      }
-    } catch (err: unknown) {
-      self_destruct(err);
-    }
-  }
-
-  events.on('project_update', project_update_cb);
-  events.on('all_state', all_state_cb);
-  if (DEBUG_APP) {
-    console.debug('listenProject finished setting up');
-  }
-
-  return detach_cb;
-}
-
 /**
  * Returns the current Data PouchDB of a project. This waits for the initial
  * sync to finish enough to know if the project exists or not before returning
@@ -362,7 +106,7 @@ export async function getDataDB(
 ): Promise<PouchDB.Database<ProjectDataObject>> {
   // Wait for all_projects_updated to possibly change before returning
   // error/data DB if it's ready.
-  await waitForStateOnce(() => all_projects_updated);
+  //await waitForStateOnce(() => all_projects_updated);
   if (active_id in data_dbs) {
     return data_dbs[active_id].local;
   } else {
@@ -443,7 +187,7 @@ export async function getProjectDB(
 ): Promise<PouchDB.Database<ProjectMetaObject>> {
   // Wait for all_projects_updated to possibly change before returning
   // error/data DB if it's ready.
-  await waitForStateOnce(() => all_projects_updated);
+  //await waitForStateOnce(() => all_projects_updated);
   if (active_id in metadata_dbs) {
     return metadata_dbs[active_id].local;
   } else {
@@ -487,6 +231,7 @@ export function listenProjectDB(
   );
 }
 
+// Get all 'listings' (conductor server links) from the local directory database
 export async function getAllListings(): Promise<ListingsObject[]> {
   const listings: ListingsObject[] = [];
   const res = await directory_db.local.allDocs({
